% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{appmove}
\alias{appmove}
\title{add two numbers}
\usage{
appmove(e, d)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
sum of two numbers
}
\description{
The movement of the  (Description)
}
\details{
Everytime user presses a key(Details)
}
\examples{

}
\author{
Caleb
}
