% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{bg.matrix}
\alias{bg.matrix}
\title{top layer}
\usage{
bg.matrix(e)
}
\arguments{
\item{e}{the list of the display}
}
\value{
a new layer of plot, including numbers and colorful rectangulars
}
\description{
generate top layer
}
\details{
this function generates the top layer of figure, including numbers and
colorful revtangulars
}
\examples{

}
\author{
Rongzi
}
