% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{bg}
\alias{bg}
\title{Background plot}
\usage{
bg(e)
}
\arguments{
\item{e}{List of the figure with all elements}
}
\value{
updated figure
}
\description{
Generate the backgroud grid
}
\details{
The background is a 4*4 grid
}
\author{
Rongzi
}
