% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{e.init}
\alias{e.init}
\title{initialization}
\usage{
e.init()
}
\arguments{
\item{e}{the list of the display}
}
\value{
updated e
}
\description{
initialization
}
\details{
Initialize the elemennts of the figure
}
\examples{

}
\author{
Rongzi
}
