% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{fail.end}
\alias{fail.end}
\title{fail condition}
\usage{
fail.end(e)
}
\arguments{
\item{e}{the list of the display}
}
\value{
game end message
}
\description{
fail condition
}
\details{
This function checks if the game fail
}
\examples{

}
\author{
Rongzi
}
