% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{play2048}
\alias{play2048}
\title{Launch the Game}
\usage{
play2048()
}
\description{
This function will launch the Game with shiny app.
'W' will rotate the tetrominos, 'A' will move the tetrominos to left,  and 'D' will move the tetrominos to right
}
\author{
Tingting & Chang
}
