% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{appmove}
\alias{appmove}
\title{movement function}
\usage{
appmove(e, d)
}
\arguments{
\item{e}{the list of the display}

\item{d}{direction of the action}
}
\value{
updated figure
}
\description{
The movement of the figure
}
\details{
This function generate a new matrix, which is an element of figure e.
}
\examples{

}
\author{
Rongzi
}
