% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{bg}
\alias{bg}
\title{background plot}
\usage{
bg(e)
}
\arguments{
\item{e}{the list of the display}
}
\value{
sum of two numbers
}
\description{
Generate the backgroud grid (Description)
}
\details{
The background is a 4*4 grid (Details)
}
\examples{

}
\author{
Caleb
}
