% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{create}
\alias{create}
\title{creation}
\usage{
create(e)
}
\arguments{
\item{e}{the list of the display}
}
\value{
updated figure
}
\description{
create a number in a random position
}
\details{
Create a number 2 in a random position. For the beginning of the game
this function will generate '2' in two random positions
}
\examples{

}
\author{
Rongzi
}
