% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{fail.end}
\alias{fail.end}
\title{add two numbers}
\usage{
fail.end(e)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
sum of two numbers
}
\description{
Add up two numbers (Description)
}
\details{
We want to add up two numbers, blalala... (Details)
}
\examples{

}
\author{
Caleb
}
